Adds a method to generate weighted sidekiq queues
Before, wiring up the sidekiq queues was a manual process that required
a full file of logic to manage. Now, with this commit, there is a
function to generate the weighted queues ahead of time.

This new function can be used in the `sidekiq.rb` file of an
implementing service, after the service (and all it's subscriptions)
have been fully loaded. Then the sidekiq options array may be modified
(or replaced) with the results of this function.