Fixes the schedule being set by heartbeats
The previous heartbeat schedule was misusing the scheduler API in such
a way that it was firing heartbeats multiple times when there was more
than one active sidekiq process. This corrects the setup to better
respect the dynamic setting and to use the proper `update_schedule` call
to load the just set schedule.