Uses cron format to schedule heartbeat
The "every: 1min" format was causing multiple heartbeats to fire in the same
minute if there were multiple sidekiq processes with the dynamic setting
turned off (which is default). Switching this to the cron syntax helps
synchronize all the sidekiq-scheduler processes to the same instant (so
long as ntp is up to date and in sync).