Test against a real redis database (#29)
* Test against a real redis database

Because Sidekiq no longer allows injection of a connection pool after
version 6, continuing to use the FakeRedis gem in the test suite is not
practical. Docker has made this sort of thing truly trivial in the time
since this was first set up however, so go with the simple solution and
run the test suite in an environment that requires substantially less
mocking and hand-waving.

* Changelog