Force-reload the SidekiqScheduler schedule
When running with dynamic:false, SidekiqScheduler supposedly allows a
manual reload of the scheduler after a `set_schedule` with a single
reload command; but that command just re-applies the schedule currently
in memory, not the schedule in Redis modified by `set_schedule`. Load
the schedule from Redis into memory before re-applying it to get the
heartbeat to actually initialize properly.